chore: cleanup seo_sidebar experiment - it won
