+++ packages/shared/src/components/sidebar/index.ts
@@ -1,8 +1,7 @@
 export * from './SquadSection';
 export * from './DiscoverSection';
-export * from './ContributeSection';
 export * from './MobileMenuIcon';
 export * from './SidebarUserButton';
 export * from './MyFeedButton';
 export * from './SidebarBottomSection';
-export * from './ManageSection';
+export * from './ActivitySection';
